Extend prebuild script to work for vscode too
Tekton and Registry commands are now deleted from VS Code extension in the jenkins build
